Creacion de las animaciones

+++ Rootbound/Assets/ScriptEnemigo/InformacionEnemigo.cs
@@ -11,7 +11,7 @@ public enum tiposDeEnemigo
 public class InformacionEnemigo : ScriptableObject
 {
     [SerializeField] private string nombre;
-    [SerializeField] private int vida;
+    [SerializeField] private float vida;
     [SerializeField] private tiposDeEnemigo tipoDeEnemigo;
     [SerializeField] private float daño;
     [SerializeField] private float velocidadMovimiento;
@@ -19,7 +19,7 @@ public class InformacionEnemigo : ScriptableObject
     [SerializeField] private GameObject modelo;
 
     public string Nombre { get { return nombre; }}
-    public int Vida { get { return vida; } }
+    public float Vida { get { return vida; } }
     public tiposDeEnemigo TipoDeEnemigo { get { return tipoDeEnemigo; } }
     public float Daño { get { return daño; } }
     public float VelocidadMovimiento { get { return velocidadMovimiento; } }
